refactor(gen-schema-view): simplify genkit flow
